% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_s4_classes.R
\name{assembleLinearModels}
\alias{assembleLinearModels}
\title{assembleLinearModels}
\arguments{
\item{theObject}{an instance of MediateModelVariables}
}
\value{
an instance of MediateLinearModels with forward and reverse linear models created using the variable vectors in theObject
}
\description{
create forward and reverse(Y and M are swapped) linear models for use with mediation analysis
}
\author{
Michael Gooch
}
