% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_s4_classes.R
\name{generateData}
\alias{generateData}
\title{generateData}
\arguments{
\item{theObject}{the object to be used in generation of the data, should be an instance of \code{MediateDataGenerationParameters}}

\item{bM.ind}{index for the betaM vector stored in theObject}

\item{SEED}{SEED to use with the resulting data}
}
\value{
an instance of \code{MediateModelVariables} with random values determined by current rng state, MediateDataGenerationParameters, and bM.ind,
}
\description{
use data generation parameters with the random number generator to generate simulated data
}
\author{
Michael Gooch
}
