% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_s4_classes.R
\name{generateDataMatrix}
\alias{generateDataMatrix}
\title{generateDataMatrix}
\arguments{
\item{theObject}{an instance of MediateDataGenerationParameters}

\item{initial_SEED}{the SEED that will be used with set.seed before generating the data, the generated MediateModelVariables will be provided incremented SEED values starting with initial_SEED + 1}
}
\value{
matrix with dimensions of \code{rows = length(theObject@betaM)} and \code{columns = theObject@nSim} containing the generated data
}
\description{
generate all the data vectors and rng seeds needed to do mediation with parallel processing
}
\author{
Michael Gooch
}
