% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_process_mediate.R
\name{mediate_parallel}
\alias{mediate_parallel}
\title{Perform Mediation Using Parallel Processing}
\usage{
mediate_parallel(list_of_job_args, nSimImai = 1000, use_cpp = F,
  num_jobs = getOption("mediate.jobs", parallel::detectCores() - 1))
}
\arguments{
\item{list_of_job_args}{A list of \code{MediateModelVariables} objects}

\item{nSimImai}{number of mediation sims to perform}
}
\value{
list of \code{MediationProbValues} objects
}
\description{
starts up multiple processes, configures their environment, and runs simulation and mediation analysis for each item in the input
}
