% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_s4_classes.R
\docType{class}
\name{MediateDataGenerationParameters}
\alias{MediateDataGenerationParameters}
\title{MediateDataGenerationParameters}
\description{
Convenience class for use with other methods and classes to generate mediation data for use with parallel processing.
}
\section{Slots}{

\describe{
\item{\code{n}}{is the sample size.}

\item{\code{pX}}{is the minor allele frequency}

\item{\code{gamma0}}{is the intercept for M}

\item{\code{gammaX}}{is the association of X with M}

\item{\code{varM}}{is the variance of M}

\item{\code{beta0}}{is the intercept for Y}

\item{\code{betaX}}{is the direct effect of X on Y}

\item{\code{betaM}}{is a vector of different associations of M with Y}

\item{\code{varY}}{is the variance of Y}

\item{\code{nSim}}{is the number of simulations to run}
}}

\author{
Michael Gooch
}
