% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_s4_classes.R
\docType{class}
\name{MediateLinearModels}
\alias{MediateLinearModels}
\title{MediateLinearModels}
\description{
convenience container for 4 linear models, 2 pairs, intended for forward and reverse mediation analysis
}
\section{Slots}{

\describe{
\item{\code{med.fit}}{linear model for the mediator}

\item{\code{out.fit}}{linear model for the outcome}

\item{\code{med.fit.r}}{linear model for the mediator with M and Y swapped}

\item{\code{out.fit.r}}{linear model for the outcome with M and Y swapped}
}}

\author{
Michael Gooch
}
