% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_s4_classes.R
\docType{class}
\name{MediateModelVariables}
\alias{MediateModelVariables}
\title{MediateModelVariables}
\description{
convenience container for vectors of simulated data to be used with other methods and classes to generate linear models for use with mediation analysis in parallel processes
}
\section{Slots}{

\describe{
\item{\code{X}}{is the exposure}

\item{\code{M}}{is the mediator}

\item{\code{Y}}{is the outcome}

\item{\code{SEED}}{is the seed to use when performing mediation on the models generated from this data}
}}

\author{
Michael Gooch
}
