% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_s4_classes.R
\docType{class}
\name{MediationProbValues}
\alias{MediationProbValues}
\title{MediationProbValues}
\description{
Convenience container for the result of a forward and reverse pair of mediation analysis results
}
\section{Slots}{

\describe{
\item{\code{pval_direct}}{direct probability from forward mediation analysis}

\item{\code{pval_indirect}}{indirect probability from forward mediation analysis}

\item{\code{pval_direct_r}}{direct probability from reverse mediation analysis}

\item{\code{pval_indirect_r}}{indirect probability from reverse mediation analysis}
}}

\author{
Michael Gooch
}
