% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Umediation.R
\name{Umediation}
\alias{Umediation}
\title{Umediation}
\usage{
Umediation(n = 100, Atype = "D", Mtype = "C", Ytype = "C",
  Ctype = "C", Utype = "C", interact = FALSE, muC = 0, varC = 1,
  muU = 0, varU = 1, gamma0 = 0, gammaC = 0, gammaU = 0,
  varA = 1, alpha0 = 0, alphaA = 0, alphaC = 0, alphaU = 0,
  varM = 1, beta0 = 0, betaA = 0, betaM = 0, betaI = 0,
  betaC = 0, betaU = 0, varY = 1, alpha = 0.05, nSim = 250,
  nBoot = 400, seed = 1, atreat = 1, acontrol = 0, use_cpp = F,
  num_cores = getOption("mediate.threads", default = 1))
}
\arguments{
\item{n}{n}

\item{Atype}{Atype}

\item{Mtype}{Mtype}

\item{Ytype}{Ytype}

\item{Ctype}{Ctype}

\item{Utype}{Utype}

\item{interact}{interact}

\item{muC}{muC}

\item{varC}{varC}

\item{muU}{muU}

\item{varU}{varU}

\item{gamma0}{gamma0}

\item{gammaC}{gammaC}

\item{gammaU}{gammaU}

\item{varA}{varA}

\item{alpha0}{alpha0}

\item{alphaA}{alphaA}

\item{alphaC}{alphaC}

\item{alphaU}{alphaU}

\item{varM}{varM}

\item{beta0}{beta0}

\item{betaA}{betaA}

\item{betaM}{betaM}

\item{betaI}{betaI}

\item{betaC}{betaC}

\item{betaU}{betaU}

\item{varY}{varY}

\item{alpha}{alpha}

\item{nSim}{nSim}

\item{nBoot}{nBoot}

\item{seed}{seed}

\item{atreat}{atreat}

\item{acontrol}{acontrol}

\item{use_cpp}{use_cpp}

\item{num_cores}{num_cores}
}
\description{
Umediation
}
\author{
Sharon Lutz, Michael Gooch
}
